requirements.txt
Flask==1.1.4
flask-cors==3.0.10
python-dotenv==0.19.0
paramiko==2.7.0
requests==2.25.1
gunicorn==21.2.0
